Pulse: adapt mixer names / volume regulation
Show sink device description instead of the sink name.
Show static string "PulseStream" instead of input info (=music).
(On other option would be the description, that is the  packagename,
Music on Console and does not fit in the bar).
If no stream is connect, show the default sink device as mixer and regulation.